Got EDIT (Update) on all story elements

+++ bedtime_story_prompter_project/bspapp/views/story_elements.py
@@ -30,7 +30,6 @@ def story_elements(request, hero_id=None, villain_id=None, setting_id=None, chal
             return redirect(reverse('story_elements'))
 
         if ("actual_method" in form_data and form_data["actual_method"] == "PUT") and hero_id is not None:
-            print("----------------FORM DATA AFTER CLICKING UPDATE", form_data)
             hero = Hero.objects.get(pk=hero_id)
             hero.name = form_data["hero"]
             hero.save()
@@ -41,9 +40,21 @@ def story_elements(request, hero_id=None, villain_id=None, setting_id=None, chal
             villain.delete()
             return redirect(reverse('story_elements'))
 
+        if ("actual_method" in form_data and form_data["actual_method"] == "PUT") and villain_id is not None:
+            villain = Villain.objects.get(pk=villain_id)
+            villain.name = form_data["villain"]
+            villain.save()
+            return redirect(reverse('story_elements'))
+
         if ("actual_method" in form_data and form_data["actual_method"] == "DELETE") and setting_id is not None:
             setting = StorySetting.objects.get(pk=setting_id)
             setting.delete()
+            return redirect(reverse('story_elements'))
+
+        if ("actual_method" in form_data and form_data["actual_method"] == "PUT") and setting_id is not None:
+            setting = StorySetting.objects.get(pk=setting_id)
+            setting.description = form_data["setting"]
+            setting.save()
             return redirect(reverse('story_elements'))
 
         if ("actual_method" in form_data and form_data["actual_method"] == "DELETE") and challenge_id is not None:
@@ -51,9 +62,21 @@ def story_elements(request, hero_id=None, villain_id=None, setting_id=None, chal
             challenge.delete()
             return redirect(reverse('story_elements'))
 
+        if ("actual_method" in form_data and form_data["actual_method"] == "PUT") and challenge_id is not None:
+            challenge = Challenge.objects.get(pk=challenge_id)
+            challenge.description = form_data["challenge"]
+            challenge.save()
+            return redirect(reverse('story_elements'))
+
         if ("actual_method" in form_data and form_data["actual_method"] == "DELETE") and template_id is not None:
             template = StoryTemplate.objects.get(pk=template_id)
             template.delete()
+            return redirect(reverse('story_elements'))
+
+        if ("actual_method" in form_data and form_data["actual_method"] == "PUT") and template_id is not None:
+            template = StoryTemplate.objects.get(pk=template_id)
+            template.template = form_data["template"]
+            template.save()
             return redirect(reverse('story_elements'))
 
         else:
